feat: Update batch prediction and configuration management
- Changed 'predicted_days' validation from integer to numeric in BatchController.
- Updated migration to change 'predicted_days' column type from integer to double.
- Enhanced PredictionPage to handle decimal predictions and added latency tracking.
- Improved ParameterComponents to lock fields during germination detection and added batch selection.
- Integrated Random Forest model for germination prediction in the Python API.
- Added new API endpoint for predicting germination days based on environmental parameters.
- Introduced a new model file 'germination_model.pkl' for machine learning predictions.

+++ backend/database/migrations/2026_03_28_072838_create_batches_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('batch_id')->unique(); // Halimbawa: B-2026-001
             $table->date('date_planted');
-            $table->integer('predicted_days'); // Ilang araw ang prediction ng AI
+            $table->double('predicted_days'); // Ilang araw ang prediction ng AI
             $table->date('actual_germination_date')->nullable(); // Nullable habang wala pang detection
             $table->timestamps();
         });
